Rip out all the DOM mutation stuff.
In addition to not working in IE, it doesn't fire from a node being removed in Firefox or Safari.
Replace that with adding explicit listeners for the content div being replaced, which requires
manually firing that event on a swap.

Expose the ListenerList from live and remove listeners from the list if they return false on a fire
to make that easier.